+++ src/main/java/roomescape/application/dto/ReservationServiceResponse.java
@@ -2,7 +2,7 @@ package roomescape.application.dto;
 
 import java.time.LocalDate;
 import java.util.List;
-import roomescape.domain.Reservation;
+import roomescape.domain.entity.Reservation;
 
 public record ReservationServiceResponse(
         long id,
