+++ src/main/java/roomescape/domain/repository/MemberRepository.java
@@ -3,7 +3,7 @@ package roomescape.domain.repository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Member;
+import roomescape.domain.entity.Member;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
