+++ src/main/java/roomescape/domain/repository/ThemeRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import roomescape.domain.Theme;
+import roomescape.domain.entity.Theme;
 
 public interface ThemeRepository extends JpaRepository<Theme, Long> {
