+++ src/main/java/roomescape/domain/repository/TimeRepository.java
@@ -1,7 +1,7 @@
 package roomescape.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.entity.ReservationTime;
 
 public interface TimeRepository extends JpaRepository<ReservationTime, Long> {
 }
