+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -11,6 +11,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import roomescape.domain.entity.Reservation;
+import roomescape.domain.entity.ReservationTime;
 
 class ReservationTest {
