+++ src/test/java/roomescape/presentation/controller/admin/AdminReservationControllerIntTest.java
@@ -19,8 +19,8 @@ import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.application.auth.dto.MemberIdDto;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Theme;
+import roomescape.domain.entity.ReservationTime;
+import roomescape.domain.entity.Theme;
 import roomescape.domain.repository.ReservationRepository;
 import roomescape.infrastructure.jwt.JwtTokenProvider;
 import roomescape.testFixture.JdbcHelper;
